added: authorization

+++ src/api/news/news.controller.ts
@@ -6,18 +6,25 @@ import {
   Param,
   Patch,
   Post,
+  UseGuards,
 } from '@nestjs/common';
 import { CreateNewsDto } from './dto/create-news.dto';
 import { UpdateNewsDto } from './dto/update-news.dto';
 import { NewsService } from './news.service';
+import { AuthGuard } from '../auth/auth.guard';
+import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
+import { GetUserId } from '../../shared/decorators/get-user-id.decorator';
 
+@ApiTags('news')
 @Controller('news')
 export class NewsController {
   constructor(private readonly newsService: NewsService) {}
 
+  @UseGuards(AuthGuard)
+  @ApiBearerAuth()
   @Post()
-  create(@Body() createNewsDto: CreateNewsDto) {
-    return this.newsService.create({ ...createNewsDto, userId: 4 });
+  create(@Body() createNewsDto: CreateNewsDto, @GetUserId() userId: number) {
+    return this.newsService.create({ ...createNewsDto, userId });
   }
 
   @Get()
@@ -30,16 +37,25 @@ export class NewsController {
     return this.newsService.findOne(Number(id));
   }
 
+  @UseGuards(AuthGuard)
+  @ApiBearerAuth()
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateNewsDto: UpdateNewsDto) {
+  update(
+    @Param('id') id: string,
+    @Body() updateNewsDto: UpdateNewsDto,
+    @GetUserId() userId: number,
+  ) {
     return this.newsService.update({
       id: Number(id),
+      userId,
       updateDto: updateNewsDto,
     });
   }
 
+  @UseGuards(AuthGuard)
+  @ApiBearerAuth()
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.newsService.remove(Number(id));
+  remove(@Param('id') id: string, @GetUserId() userId: number) {
+    return this.newsService.remove({ id: Number(id), userId });
   }
 }
